:fist: prevent from saving uploaded geojson
