+++ geoinfo/management/commands/initiate_db.py
@@ -4,7 +4,7 @@ from django.core.management.base import BaseCommand
 from django.core.management import call_command
 from django.conf import settings
 
-from utils.common import get_geojson_file
+# from utils.common import get_geojson_file
 from utils.geoparser import GeoJSONParser
 
 
@@ -28,7 +28,7 @@ class Command(BaseCommand):
         geofilenames = os.listdir(settings.INIT_GEOJSON_FOLDER)
         fullpathes = [os.path.join(
                       settings.INIT_GEOJSON_FOLDER, x) for x in geofilenames]
-        geojsons = list(map(get_geojson_file, fullpathes))
+        geojsons = list(map(GeoJSONParser.get_geojson_file, fullpathes))
         geojsons.sort(key=lambda x: x['ctracker_config']['AL'])
         list(map(GeoJSONParser.geojson_to_db, geojsons))
