+++ init_geo_data/AL0.geojson
@@ -4,7 +4,7 @@
 "ctracker_config" :  {"ZOOM":6, "AL":0},                                                                                
 "features": [
 { "type": "Feature", 
-"properties": { "ADDRESS": "Ukraine", "ID": "root", "PARENT": "", "ORG_NAMES": null, "ORG_TYPES": null, "CENTROID": "36.2543045,50.0075170" }, 
+"properties": { "ADDRESS": "Ukraine", "ID": "root", "PARENT": "", "CENTROID": "36.2543045,50.0075170" }, 
 "geometry": { "type": "Polygon", "coordinates": [ [ [ 36.254347282380181, 50.007771170054653 ], [ 36.254542740692763, 50.007721737070284 ], [ 36.254261806847659, 50.007262920920837 ], [ 36.254066348535076, 50.007312354377092 ], [ 36.254347282380181, 50.007771170054653 ] ] ] } 
 }]
 }
