+++ init_geo_data/AL4_CNAP_prosec.geojson
@@ -1,7 +1,7 @@
 {
 "type": "FeatureCollection",
 "crs": { "type": "name", "properties": { "name": "urn:ogc:def:crs:OGC:1.3:CRS84" } },
-"ctracker_config" :  {"ZOOM":17, "AL":4},
+"ctracker_config" :  {"ZOOM":17, "AL":4, "ORG_TYPES": {"prosecutors": "Прокуратура", "CNAP": "ЦПАУ"}},
 "features": [
 { "type": "Feature", "properties": { "ADDRESS": "61024, м. Харків, вул.Студентська, 5\nтел. 704-15-15, e-mail gorod@prok.kh.ua", "ID": "21citkyivs0001", "PARENT": "21citkyivs", "ORG_NAMES": "Прокуратура м. Харкова", "ORG_TYPES": "prosecutors", "CENTROID": "36.2543045,50.0075170" }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 36.254347282380181, 50.007771170054653 ], [ 36.254542740692763, 50.007721737070284 ], [ 36.254261806847659, 50.007262920920837 ], [ 36.254066348535076, 50.007312354377092 ], [ 36.254347282380181, 50.007771170054653 ] ] ] } },
 { "type": "Feature", "properties": { "ADDRESS": "61072, вул.Тобольська, 55-а тел. 773-04-69, e-mail dzerzh@prok.kh.ua", "ID": "21citdzerz0001", "PARENT": "21citdzerz", "ORG_NAMES": "Прокуратура Дзержинського району м. Харкова", "ORG_TYPES": "prosecutors", "CENTROID": "36.2239192,50.0304025" }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 36.223543772302719, 50.030430551788598 ], [ 36.224260352065663, 50.030507271224472 ], [ 36.224294767241197, 50.030374625393613 ], [ 36.223578187478246, 50.03029790574584 ], [ 36.223543772302719, 50.030430551788598 ] ] ] } },
